requirements.txt
imutils==0.5.2
opencv_python==4.0.0.21
Pillow==6.0.0
tkinter==8.6
